requirements.txt
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
gym-retro>=0.8.0
